+++ src/sled/filter.rs
@@ -4,7 +4,7 @@ use crate::{color::Rgb, led::Led, sled::Sled};
 
 #[derive(Clone)]
 pub struct Filter {
-    led_indices: HashSet<usize>,
+    led_indices: HashSet<u16>,
 }
 
 impl From<&[Led]> for Filter {
@@ -17,8 +17,8 @@ impl From<&[Led]> for Filter {
     }
 }
 
-impl From<HashSet<usize>> for Filter {
-    fn from(value: HashSet<usize>) -> Self {
+impl From<HashSet<u16>> for Filter {
+    fn from(value: HashSet<u16>) -> Self {
         Filter { led_indices: value }
     }
 }
@@ -60,8 +60,8 @@ impl Filter {
 }
 
 impl IntoIterator for Filter {
-    type Item = usize;
-    type IntoIter = hash_set::IntoIter<usize>;
+    type Item = u16;
+    type IntoIter = hash_set::IntoIter<u16>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.led_indices.into_iter()
@@ -69,7 +69,7 @@ impl IntoIterator for Filter {
 }
 
 impl IntoIterator for &Filter {
-    type Item = usize;
+    type Item = u16;
     type IntoIter = hash_set::IntoIter<Self::Item>;
 
     fn into_iter(self) -> Self::IntoIter {
@@ -81,20 +81,20 @@ impl IntoIterator for &Filter {
 impl Sled {
     pub fn set_filter(&mut self, filter: &Filter, color: Rgb) {
         for i in filter {
-            self.leds[i].color = color;
+            self.leds[i as usize].color = color;
         }
     }
 
     pub fn modulate_filter<F: Fn(&Led) -> Rgb>(&mut self, filter: &Filter, color_rule: F) {
         for i in filter {
-            let led = &mut self.leds[i];
+            let led = &mut self.leds[i as usize];
             led.color = color_rule(led)
         }
     }
 
     pub fn for_each_in_filter<F: FnMut(&mut Led)>(&mut self, filter: &Filter, mut func: F) {
         for i in filter {
-            let led = &mut self.leds[i];
+            let led = &mut self.leds[i as usize];
             func(led);
         }
     }
